Update esg.md
Added new policies on Competition and Anti-Harassment. Updated links and text in Anti-Bribery section to reflect the 2023 update of OECD Guidelines. Added new sections to the intro text at the beginning. Fixed link in Environmental section to point to new version of Luonnonsuojelulaki. Fixed links in Anti-Discrimination to point to Yhdenvertaisuusvaltuutettu pages instead of obsoleted syrjinta.fi.